requirements.txt
coverage==4.5.3
Pillow==5.4.1
requests==2.21.0
schwifty==2018.9.1
